# feat: allow to select a subset of columns