--- a/Cargo.toml
+++ b/Cargo.toml
@@ -28,3 +28,6 @@ rstest = { version = "0.18.2", default-features = false }
 [features]
 extension-module = ["pyo3/extension-module"]
 default = ["extension-module"]
+# feature for tests only. This makes Python::with_gil auto-initialize Python
+# interpreters, which allows us ot instantiate Python objects in tests
+tests = ["pyo3/auto-initialize"]