--- a/Makefile
+++ b/Makefile
@@ -9,7 +9,7 @@ pytest	= pytest -v
 ## Rust
 clippy		= cargo clippy
 fmt		= cargo fmt
-cargo-test	= cargo test
+cargo-test	= cargo test --no-default-features
 ## Docs
 pdoc	= pdoc -o docs python/fastexcel