--- a/Makefile
+++ b/Makefile
@@ -13,16 +13,25 @@ cargo-test	= cargo test
 ## Docs
 pdoc	= pdoc -o docs python/fastexcel
 
-lint:
+lint-python:
 	$(ruff)
 	$(format)  --check --diff
+
+lint-rust:
 	$(mypy)
 	$(clippy)
-format:
+
+lint: lint-rust lint-python
+
+format-python:
 	$(ruff) --fix
 	$(format)
+
+format-rust:
 	$(fmt)
 	$(clippy) --fix --lib -p fastexcel --allow-dirty --allow-staged
+
+format: format-rust format-python
 
 install-test-requirements:
 	pip install -U -r test-requirements.txt -r build-requirements.txt
@@ -39,9 +48,13 @@ dev-install:
 prod-install:
 	./prod_install.sh
 
-test:
+test-rust:
 	$(cargo-test)
+
+test-python:
 	$(pytest)
+
+test: test-rust test-python
 
 doc:
 	$(pdoc)