closes #172 . Depends on #186