@lukapeschke 2024-02-26
@PrettyWood I'm looking at the implem for column ranges right now, and I'm thinking about wether column names should be loaded and validated on sheet instatiation. However, this would have an impactt:
* `ExcelSheet` _instantiation_ would become slightly slower and less efficient memory-wise, as data loading & validation would occur on instantiation. 
* Errors would occur earlier: the column selection would be validated on sheet instantiation rather than on conversion
* `column_names` could become a sheet attribute
* It _might_ be possible that this would allow to make functions obtaining data by ref lazy (not sure about this one, depends if I can elide lifetimes, as they don't play well with PyO3: https://pyo3.rs/main/class.html?highlight=lifetime#no-lifetime-parameters . Maybe by making the `reader` owned ?)

Any thoughts about this ?