@PrettyWood 2024-02-26
Yeah this is what I wanted originally when I said "Would be nice to add to ExcelSheet the range of indices and list of columns. Currently we only have the width"
I'd rather have a slight drop in performance-wise but better API and validation.
In my mind once you load the excelsheet you could have the range of column indices and names.
And if you select some they are directly validated at instantiation with better error messages (even with fuzzysearch to say "did you mean this column?" which is clearly not needed but why not haha)
Could we try to go with this approach and see the perf impact?